\name{PARCOR-package}
\alias{PARCOR-package}
\alias{PARCOR}
\docType{package}
\title{
\packageTitle{PARCOR}
}
\description{
\packageDescription{PARCOR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PARCOR}
\packageIndices{PARCOR}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{PARCOR}

Maintainer: \packageMaintainer{PARCOR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
